# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Govershop API
  description: |
    # Govershop Top-up API
    
    API untuk aplikasi top-up produk digital (pulsa, paket data, game voucher, dll).
    
    ## Integrasi
    - **Digiflazz** - Supplier produk digital
    - **Pakasir** - Payment gateway (QRIS, Virtual Account)
    
    ## Flow Transaksi
    1. User browse produk (`GET /products`)
    2. User buat order (`POST /orders`)
    3. User pilih metode pembayaran (`POST /orders/{id}/pay`)
    4. User bayar via QRIS/VA
    5. Webhook konfirmasi pembayaran
    6. Sistem proses topup otomatis
    7. User cek status (`GET /orders/{id}/status`)
    
  version: 1.0.0
  contact:
    name: Govershop Support
    email: support@govershop.com
servers:
  - url: http://localhost:8080
    description: Development Server (Local)
  - url: https://govershop-be-production.up.railway.app
    description: Production Server (Railway)
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Daftar Produk
      description: |
        Mengambil daftar semua produk yang tersedia.
        Bisa difilter berdasarkan category atau brand.
      parameters:
        - name: category
          in: query
          description: Filter berdasarkan kategori (e.g., Pulsa, Data, Games)
          required: false
          schema:
            type: string
            example: Pulsa
        - name: brand
          in: query
          description: Filter berdasarkan brand (e.g., TELKOMSEL, XL, MOBILE LEGENDS)
          required: false
          schema:
            type: string
            example: TELKOMSEL
      responses:
        '200':
          description: Daftar produk berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      total:
                        type: integer
                        example: 14
...
  /api/v1/products/brands:
    get:
      tags:
        - Products
      summary: Daftar Brand
      description: Mengambil daftar brand produk yang tersedia
      parameters:
        - name: category
          in: query
          description: Filter brand berdasarkan kategori
          required: false
          schema:
            type: string
            example: Pulsa
      responses:
        '200':
          description: Daftar brand berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      brands:
                        type: array
                        items:
                          $ref: '#/components/schemas/Brand'
  /api/v1/products/{sku}:
    get:
      tags:
        - Products
      summary: Detail Produk
      description: Mengambil detail produk berdasarkan SKU
      parameters:
        - name: sku
          in: path
          description: Kode SKU produk (buyer_sku_code)
          required: true
          schema:
            type: string
            example: tsel5
      responses:
        '200':
          description: Detail produk berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Produk tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # VALIDATION
  # ==========================================
  /api/v1/validate-account:
    post:
      tags:
        - Validation
      summary: Validasi Akun Game
      description: |
        Memvalidasi akun game sebelum topup untuk memastikan ID benar.
        Menggunakan produk "Check Username" dari Digiflazz.
        **Biaya validasi akan dikenakan** (termasuk dalam total harga).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - brand
                - customer_no
              properties:
                brand:
                  type: string
                  description: Brand/nama game (e.g., MOBILE LEGENDS, FREE FIRE)
                  example: MOBILE LEGENDS
                customer_no:
                  type: string
                  description: User ID + Zone ID (untuk ML), atau Player ID (untuk game lain)
                  example: "123456781234"
            example:
              brand: MOBILE LEGENDS
              customer_no: "123456781234"
      responses:
        '200':
          description: Validasi berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Validasi berhasil
                  data:
                    type: object
                    properties:
                      is_valid:
                        type: boolean
                        description: Apakah akun valid
                        example: true
                      account_name:
                        type: string
                        description: Nama akun (IGN)
                        example: "PLAYERXYZ"
                      customer_no:
                        type: string
                        example: "123456781234"
                      brand:
                        type: string
                        example: MOBILE LEGENDS
                      message:
                        type: string
                        example: Transaksi Sukses
                      validation_fee:
                        type: number
                        description: Biaya validasi yang dikenakan
                        example: 9
        '400':
          description: Request tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produk validasi tidak ditemukan untuk brand ini
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/calculate-price:
    post:
      tags:
        - Validation
      summary: Kalkulasi Harga Total
      description: |
        Menghitung total harga yang harus dibayar termasuk:
        - Harga produk
        - Biaya admin (check username)
        
        **Biaya payment gateway tidak dihitung di sini** (ditanggung/dihitung oleh gateway).
        **Gunakan endpoint ini sebelum checkout** untuk menampilkan breakdown harga.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - buyer_sku_code
                - payment_method
                - brand
              properties:
                buyer_sku_code:
                  type: string
                  description: SKU produk yang akan dibeli
                  example: pre29536099
                payment_method:
                  type: string
                  description: Metode pembayaran
                  enum: [qris, bni_va, bri_va, mandiri_va, permata_va, cimb_niaga_va]
                  example: qris
                brand:
                  type: string
                  description: Brand untuk check username fee
                  example: MOBILE LEGENDS
            example:
              buyer_sku_code: pre29536099
              payment_method: qris
              brand: MOBILE LEGENDS
      responses:
        '200':
          description: Kalkulasi berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Kalkulasi harga berhasil
                  data:
                    type: object
                    properties:
                      product_price:
                        type: number
                        description: Harga produk
                        example: 38527
                      admin_fee:
                        type: number
                        description: Biaya Admin (Flat)
                        example: 10
                      payment_fee:
                        type: number
                        description: Biaya payment gateway (ditanggung user via gateway)
                        example: 0
                      total_price:
                        type: number
                        description: Total yang harus dibayar (produk + admin fee)
                        example: 38537
                      product_name:
                        type: string
                        example: "Mobile Legends 144 Diamonds (Indonesia)"
                      payment_method_label:
                        type: string
                        example: QRIS
                      breakdown:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                label:
                                  type: string
                                  example: "Mobile Legends 144 Diamonds (Indonesia)"
                                amount:
                                  type: number
                                  example: 38527
                            example:
                              - label: "Mobile Legends 144 Diamonds (Indonesia)"
                                amount: 38527
                              - label: "Biaya Admin"
                                amount: 10
        '400':
          description: Request tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produk tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================
  # ORDERS
  # ==========================================
  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Buat Order Baru
      description: |
        Membuat order baru untuk pembelian produk digital.
        Order akan dibuat dengan status `pending` dan menunggu pembayaran.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - buyer_sku_code
                - customer_no
              properties:
                buyer_sku_code:
                  type: string
                  description: Kode SKU produk yang akan dibeli
                  example: tsel5
                customer_no:
                  type: string
                  description: Nomor tujuan topup (nomor HP, ID game, dll)
                  example: "081234567890"
                customer_email:
                  type: string
                  description: Email customer (opsional, untuk notifikasi)
                  example: customer@email.com
                customer_phone:
                  type: string
                  description: Nomor HP customer (opsional, untuk notifikasi)
                  example: "081234567890"
                customer_name:
                  type: string
                  description: Nama customer (opsional)
                  example: John Doe
            example:
              buyer_sku_code: tsel5
              customer_no: "081234567890"
              customer_email: customer@email.com
              customer_phone: "081234567890"
              customer_name: John Doe
      responses:
        '201':
          description: Order berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order berhasil dibuat
                  data:
                    $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Request tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produk tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/track:
    get:
      tags:
        - Orders
      summary: Track Order
      description: Track order history by phone number
      parameters:
        - name: phone
          in: query
          description: Phone number used for orders
          required: true
          schema:
            type: string
            example: "081234567890"
      responses:
        '200':
          description: List order berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Phone number required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Detail Order
      description: Mengambil detail order berdasarkan ID
      parameters:
        - name: id
          in: path
          description: UUID Order ID
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Detail order berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{id}/pay:
    post:
      tags:
        - Orders
        - Payment
      summary: Initiate Payment
      description: |
        Memulai proses pembayaran untuk order.
        Akan mengembalikan QR code (untuk QRIS) atau nomor VA.
      parameters:
        - name: id
          in: path
          description: UUID Order ID
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_method
              properties:
                payment_method:
                  type: string
                  description: Metode pembayaran yang dipilih
                  enum:
                    - qris
                    - bni_va
                    - bri_va
                    - mandiri_va
                    - permata_va
                    - cimb_niaga_va
                  example: qris
            example:
              payment_method: qris
      responses:
        '200':
          description: Payment berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Pembayaran berhasil dibuat
                  data:
                    $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Request tidak valid atau order sudah dibayar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel Order
      description: |
        Membatalkan order yang masih pending atau waiting_payment.
        Order yang sudah dibayar tidak bisa dibatalkan.
      parameters:
        - name: id
          in: path
          description: UUID Order ID
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Order berhasil dibatalkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order berhasil dibatalkan
        '400':
          description: Order tidak dapat dibatalkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{id}/status:
    get:
      tags:
        - Orders
      summary: Cek Status Order
      description: Mengecek status order dan pembayaran
      parameters:
        - name: id
          in: path
          description: UUID Order ID
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Status order berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      order_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [pending, waiting_payment, paid, processing, success, failed, cancelled, refunded]
                        example: success
                      status_label:
                        type: string
                        example: Berhasil
                      serial_number:
                        type: string
                        description: Serial number dari provider (jika berhasil)
                        example: "1234567890123456"
                      message:
                        type: string
                        example: Transaksi berhasil
                      payment:
                        $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Order tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/products/{sku}/image:
    put:
      tags:
        - Admin
      summary: Set Image Product
      description: Set or update product/brand image
      parameters:
        - name: sku
          in: path
          description: Buyer SKU Code
          required: true
          schema:
            type: string
        - name: image_url
          in: query
          description: URL of the image
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Product image updated successfully"
    delete:
      tags:
        - Admin
      summary: Delete Image Product
      description: Remove product/brand image
      parameters:
        - name: sku
          in: path
          description: Buyer SKU Code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Product image deleted successfully"
                    example: true
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderResponse'
                      total:
                        type: integer
                        example: 100
...
  /api/v1/admin/products:
    get:
      tags:
        - Admin
      summary: Daftar Produk (Admin)
      description: |
        Mengambil daftar semua produk termasuk yang unavailable.
        Digunakan untuk admin management.
      parameters:
        - name: limit
          in: query
          description: Jumlah produk per halaman
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Offset untuk pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Daftar produk berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminProduct'
                      total:
                        type: integer
                        example: 1000
                      limit:
                        type: integer
                        example: 50
                      offset:
                        type: integer
                        example: 0
...
  /api/v1/admin/products/{sku}/tags:
    post:
      tags:
        - Admin
      summary: Tambah Tag ke Produk
      description: Menambahkan tag baru ke produk
      parameters:
        - name: sku
          in: path
          required: true
          description: SKU produk (buyer_sku_code)
          schema:
            type: string
            example: ml5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tag
              properties:
                tag:
                  type: string
                  description: Tag yang akan ditambahkan
                  example: diamond
            example:
              tag: diamond
      responses:
        '200':
          description: Tag berhasil ditambahkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Tag berhasil ditambahkan
  /api/v1/admin/products/{sku}/tags/{tag}:
    delete:
      tags:
        - Admin
      summary: Hapus Tag dari Produk
      description: Menghapus tag dari produk
      parameters:
        - name: sku
          in: path
          required: true
          description: SKU produk (buyer_sku_code)
          schema:
            type: string
            example: ml5
        - name: tag
          in: path
          required: true
          description: Tag yang akan dihapus
          schema:
            type: string
            example: diamond
      responses:
        '200':
          description: Tag berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Tag berhasil dihapus

  # ==========================================
  # WEBHOOKS
  # ==========================================
...
components:
  schemas:
    Product:
      type: object
      properties:
        buyer_sku_code:
          type: string
          description: Kode SKU unik produk
          example: tsel5
        product_name:
          type: string
          description: Nama produk (menggunakan display_name jika tersedia)
          example: "Telkomsel 5.000"
        category:
          type: string
          description: Kategori produk
          example: Pulsa
        brand:
          type: string
          description: Brand produk
          example: TELKOMSEL
        type:
          type: string
          description: Tipe produk
          example: Umum
        price:
          type: number
          format: float
          description: Harga jual (sudah termasuk markup)
          example: 6000
        original_price:
          type: number
          format: float
          description: Harga normal (jika sedang promo)
          example: 6500
        is_available:
          type: boolean
          description: Status ketersediaan produk
          example: true
        unlimited_stock:
          type: boolean
          example: true
        stock:
          type: integer
          example: 0
        description:
          type: string
          example: "Pulsa Telkomsel Rp 5.000"
        is_promo:
          type: boolean
          description: Apakah produk sedang promo
          example: false
        is_best_seller:
          type: boolean
          description: Flag best seller
          example: false
        tags:
          type: array
          items:
            type: string
          description: Tags produk
          example: ["pulsa", "promo"]
        image_url:
          type: string
          description: URL gambar brand/produk
          example: "https://example.com/image.png"
...